fix(backend): coloca nome dos arquivos em snake_case
