+++ app/SubcategoriaCuidadoPessoal/schema_subcategoria_cuidado_pessoal.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+
+class SubcategoriaCuidadoPessoalBase(BaseModel):
+    nome: str
+    descricao: str
+
+class SubcategoriaCuidadoPessoalCreate(SubcategoriaCuidadoPessoalBase):
+    pass
+
+class SubcategoriaCuidadoPessoalRead(SubcategoriaCuidadoPessoalBase):
+    id: int
+
+    class Config:
+        orm_mode = True
