--- app/models/item_estoque.py
@@ -1,1 +1,0 @@
-# app/models/item_estoque.py
