--- app/models/restricao.py
@@ -1,9 +1,0 @@
-#app/models/restricao.py
-from sqlalchemy import Column, Integer, String
-from app.settings import Base
-
-class Restricao(Base):
-    __tablename__ = "restricao"
-
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    nome = Column(String(50), unique=True, nullable=False)
