--- app/schemas/armazenamento_schema.py
@@ -1,15 +1,0 @@
-#app/schema/armazem_schema.py
-from pydantic import BaseModel
-from typing import Optional
-
-class ArmazemBase(BaseModel):
-    local_armazem: str
-
-class ArmazemCreate(ArmazemBase):
-    pass
-
-class ArmazemRead(ArmazemBase):
-    id: int
-
-    class Config:
-        orm_mode = True
