--- app/schemas/item_estoque_schema.py
@@ -1,1 +1,0 @@
-# app/schema/item_estoque_schema.py
