--- app/schemas/restricao_schema.py
@@ -1,14 +1,0 @@
-#app/schema/restricao_schema.py
-from pydantic import BaseModel
-
-class RestricaoBase(BaseModel):
-    nome: str
-
-class RestricaoCreate(RestricaoBase):
-    pass
-
-class RestricaoRead(RestricaoBase):
-    id: int
-
-    class Config:
-        orm_mode = True
