--- app/services/item_estoque_service.py
@@ -1,1 +1,0 @@
-# app/service/item_estoque_service.py
